added protection to not delete all admin accounts

+++ app/Http/Controllers/ProfileController.php
@@ -47,8 +47,13 @@ class ProfileController extends Controller
             'password' => ['required', 'current_password'],
         ]);
 
-        User::where('id', $id)->delete();
+        if(User::where('is_admin', 1)->count() > 1){
+            User::where('id', $id)->delete();
+            return Redirect::to('/employee-accounts');
+        } else {
+            return Redirect::to('/employee-accounts?status=failed');
+        }
 
-        return Redirect::to('/employee-accounts');
+
     }
 }
